changing all int variables into float and cast -- an eventual recasting specific of the architecture will be handled by drivers

+++ src/qibocal/update.py
@@ -26,22 +26,24 @@ def replace(model: BaseModel, **update):
 def readout_frequency(freq: float, platform: Platform, qubit: QubitId):
     """Update readout frequency value in platform for specific qubit."""
     ro_channel = platform.qubits[qubit].probe
-    platform.update({f"configs.{ro_channel}.frequency": freq})
+    platform.update({f"configs.{ro_channel}.frequency": float(freq)})
 
 
 def bare_resonator_frequency(freq: float, platform: Platform, qubit: QubitId):
     """Update rbare frequency value in platform for specific qubit."""
-    platform.calibration.single_qubits[qubit].resonator.bare_frequency = int(freq)
+    platform.calibration.single_qubits[qubit].resonator.bare_frequency = float(freq)
 
 
 def dressed_resonator_frequency(freq: float, platform: Platform, qubit: QubitId):
     """Update rbare frequency value in platform for specific qubit."""
-    platform.calibration.single_qubits[qubit].resonator.dressed_frequency = int(freq)
+    platform.calibration.single_qubits[qubit].resonator.dressed_frequency = float(freq)
 
 
 def readout_amplitude(amp: float, platform: Platform, qubit: QubitId):
     """Update readout amplitude value in platform for specific qubit."""
-    platform.update({f"native_gates.single_qubit.{qubit}.MZ.0.1.probe.amplitude": amp})
+    platform.update(
+        {f"native_gates.single_qubit.{qubit}.MZ.0.1.probe.amplitude": float(amp)}
+    )
 
 
 def drive_frequency(freq: float | tuple | list, platform: Platform, qubit: QubitId):
@@ -49,7 +51,7 @@ def drive_frequency(freq: float | tuple | list, platform: Platform, qubit: Qubit
     if isinstance(freq, Iterable):
         freq = freq[0]
     drive_channel = platform.qubits[qubit].drive
-    platform.update({f"configs.{drive_channel}.frequency": freq})
+    platform.update({f"configs.{drive_channel}.frequency": float(freq)})
 
 
 def drive_amplitude(
@@ -59,9 +61,13 @@ def drive_amplitude(
     if isinstance(amp, Iterable):
         amp = amp[0]
     if rx90:
-        platform.update({f"native_gates.single_qubit.{qubit}.RX90.0.1.amplitude": amp})
+        platform.update(
+            {f"native_gates.single_qubit.{qubit}.RX90.0.1.amplitude": float(amp)}
+        )
     else:
-        platform.update({f"native_gates.single_qubit.{qubit}.RX.0.1.amplitude": amp})
+        platform.update(
+            {f"native_gates.single_qubit.{qubit}.RX.0.1.amplitude": float(amp)}
+        )
 
 
 def drive_duration(
@@ -72,11 +78,11 @@ def drive_duration(
         duration = duration[0]
     if rx90:
         platform.update(
-            {f"native_gates.single_qubit.{qubit}.RX90.0.1.duration": int(duration)}
+            {f"native_gates.single_qubit.{qubit}.RX90.0.1.duration": float(duration)}
         )
     else:
         platform.update(
-            {f"native_gates.single_qubit.{qubit}.RX.0.1.duration": int(duration)}
+            {f"native_gates.single_qubit.{qubit}.RX.0.1.duration": float(duration)}
         )
 
 
@@ -158,40 +164,56 @@ def virtual_phases(
     )
 
 
-def CZ_duration(duration: int, platform: Platform, pair: QubitPairId):
+def CZ_duration(duration: float, platform: Platform, pair: QubitPairId):
     """Update CZ duration for specific pair."""
     platform.update(
-        {f"native_gates.two_qubit.{f'{pair[0]}-{pair[1]}'}.CZ.0.1.duration": duration}
+        {
+            f"native_gates.two_qubit.{f'{pair[0]}-{pair[1]}'}.CZ.0.1.duration": float(
+                duration
+            )
+        }
     )
 
 
 def CZ_amplitude(amp: float, platform: Platform, pair: QubitPairId):
     """Update CZ amplitude for specific pair."""
     platform.update(
-        {f"native_gates.two_qubit.{f'{pair[0]}-{pair[1]}'}.CZ.0.1.amplitude": amp}
+        {
+            f"native_gates.two_qubit.{f'{pair[0]}-{pair[1]}'}.CZ.0.1.amplitude": float(
+                amp
+            )
+        }
     )
 
 
-def iSWAP_duration(duration: int, platform: Platform, pair: QubitPairId):
+def iSWAP_duration(duration: float, platform: Platform, pair: QubitPairId):
     """Update iSWAP_duration duration for specific pair."""
     platform.update(
-        {f"native_gates.two_qubit.{f'{pair[0]}-{pair[1]}'}.CZ.0.1.duration": duration}
+        {
+            f"native_gates.two_qubit.{f'{pair[0]}-{pair[1]}'}.CZ.0.1.duration": float(
+                duration
+            )
+        }
     )
 
 
 def iSWAP_amplitude(amp: float, platform: Platform, pair: QubitPairId):
     """Update iSWAP_duration amplitude for specific pair."""
     platform.update(
-        {f"native_gates.two_qubit.{f'{pair[0]}-{pair[1]}'}.CZ.0.1.amplitude": amp}
+        {
+            f"native_gates.two_qubit.{f'{pair[0]}-{pair[1]}'}.CZ.0.1.amplitude": float(
+                amp
+            )
+        }
     )
 
 
-def t1(t1: int, platform: Platform, qubit: QubitId):
+def t1(t1: float, platform: Platform, qubit: QubitId):
     """Update t1 value in platform for specific qubit."""
     platform.calibration.single_qubits[qubit].t1 = tuple(t1)
 
 
-def t2(t2: int, platform: Platform, qubit: QubitId):
+def t2(t2: float, platform: Platform, qubit: QubitId):
     """Update t2 value in platform for specific qubit."""
     platform.calibration.single_qubits[qubit].t2 = tuple(t2)
 
@@ -206,7 +228,7 @@ def drag_pulse_beta(beta: float, platform: Platform, qubit: QubitId):
     platform.update(
         {
             f"native_gates.single_qubit.{qubit}.RX.0.1.envelope.kind": "drag",
-            f"native_gates.single_qubit.{qubit}.RX.0.1.envelope.beta": beta,
+            f"native_gates.single_qubit.{qubit}.RX.0.1.envelope.beta": float(beta),
         }
     )
 
@@ -227,25 +249,25 @@ def flux_coefficients(
 
 def flux_offset(offset: float, platform: Platform, qubit: QubitId):
     """Update flux offset parameter in platform for specific qubit."""
-    platform.update({f"configs.{platform.qubits[qubit].flux}.offset": offset})
+    platform.update({f"configs.{platform.qubits[qubit].flux}.offset": float(offset)})
 
 
-def frequency_12_transition(frequency: int, platform: Platform, qubit: QubitId):
+def frequency_12_transition(frequency: float, platform: Platform, qubit: QubitId):
     channel = platform.qubits[qubit].drive_extra[1, 2]
-    platform.update({f"configs.{channel}.frequency": frequency})
-    platform.calibration.single_qubits[qubit].qubit.frequency_12 = int(frequency)
+    platform.update({f"configs.{channel}.frequency": float(frequency)})
+    platform.calibration.single_qubits[qubit].qubit.frequency_12 = float(frequency)
 
 
 def drive_12_amplitude(amplitude: float, platform: Platform, qubit: QubitId):
     platform.update(
-        {f"native_gates.single_qubit.{qubit}.RX12.0.1.amplitude": amplitude}
+        {f"native_gates.single_qubit.{qubit}.RX12.0.1.amplitude": float(amplitude)}
     )
 
 
 def drive_12_duration(duration: int | tuple | list, platform: Platform, qubit: QubitId):
     """Update drive duration value in platform for specific qubit."""
     platform.update(
-        {f"native_gates.single_qubit.{qubit}.RX12.0.1.duration": int(duration)}
+        {f"native_gates.single_qubit.{qubit}.RX12.0.1.duration": float(duration)}
     )
